refactor(interactions): migrate loot to registry
